first dev branch commit
for clarity sake changed JavaRendeEngine to be a start-up window render service application for the apps that use the rendering canvas and have the event forwarded to it.
added public interface AppHandler into JavaRendeEngine that lets the application functionality to get the forwarded events and rendering calls from the main window.
added DrawApp that contains all the drawing application functionality (which is currently not fully adapted and contains Java syntax errors and will not compile).
the main window in JavaRendeEngine with null active application will  show black screen and process key presses ESC (close), ALT-ENTER (full screen), F4-F8 (application).
