+++ src/fi/jkauppa/javarenderengine/JavaRenderEngine.java
@@ -1,17 +1,13 @@
 package fi.jkauppa.javarenderengine;
 
 import java.awt.AlphaComposite;
-import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
 import java.awt.GraphicsConfiguration;
 import java.awt.GraphicsDevice;
 import java.awt.GraphicsEnvironment;
-import java.awt.Rectangle;
-import java.awt.TexturePaint;
 import java.awt.Transparency;
-import java.awt.datatransfer.Transferable;
 import java.awt.dnd.DropTarget;
 import java.awt.dnd.DropTargetDropEvent;
 import java.awt.event.ActionEvent;
@@ -25,20 +21,11 @@ import java.awt.event.MouseListener;
 import java.awt.event.MouseMotionListener;
 import java.awt.event.MouseWheelEvent;
 import java.awt.event.MouseWheelListener;
-import java.awt.image.BufferedImage;
 import java.awt.image.VolatileImage;
-import java.io.File;
-import java.nio.file.Paths;
-
-import javax.imageio.ImageIO;
-import javax.swing.JComponent;
-import javax.swing.JFileChooser;
 import javax.swing.JFrame;
 import javax.swing.JPanel;
 import javax.swing.Timer;
-import javax.swing.TransferHandler;
 import javax.swing.UIManager;
-import javax.swing.filechooser.FileFilter;
 
 import fi.jkauppa.javarenderengine.MathLib.Direction;
 import fi.jkauppa.javarenderengine.MathLib.Matrix;
@@ -51,41 +38,15 @@ import fi.jkauppa.javarenderengine.ModelLib.Model;
 public class JavaRenderEngine extends JFrame implements KeyListener,MouseListener,MouseMotionListener,MouseWheelListener {
 	private static final long serialVersionUID = 1L;
 	private DrawApp drawapp = new DrawApp();
-	private AppHandler activeapp = drawapp;
-	private String userlocalpath = Paths.get("").toAbsolutePath().toString();
-	private String userlocaldir = System.getProperty("user.dir");
-	private String[] writeformatnames = ImageIO.getWriterFormatNames();
-	private String[] readformatnames = ImageIO.getReaderFormatNames();
+	private AppHandler activeapp = null;
 	private GraphicsEnvironment ge = GraphicsEnvironment.getLocalGraphicsEnvironment ();
 	private GraphicsDevice gd = ge.getDefaultScreenDevice ();
 	private GraphicsConfiguration gc = gd.getDefaultConfiguration ();
 	private int imagecanvaswidth = 1920;
 	private int imagecanvasheight= 1080;
-	private int resizewidth = this.imagecanvaswidth;
-	private int resizeheight = this.imagecanvasheight;
-	private boolean resizing = false;
-	private RenderPanel renderpanel = new RenderPanel(imagecanvaswidth,imagecanvasheight);
+	private boolean resizing = true;
+	private RenderPanel renderpanel = new RenderPanel();
 	private boolean windowedmode = true;
-	private Color drawcolor = Color.BLACK;
-	private float[] drawcolorhsb = {0.0f, 1.0f, 0.0f};
-	private Color erasecolor = new Color(1.0f,1.0f,1.0f,0.0f);
-	private int pencilsize = 1;
-	private int pencilshape = 1;
-	private boolean penciloverridemode = false;
-	private float penciltransparency = 1.0f;
-	private VolatileImage pencilbuffer = null;
-	private int oldpencilsize = 1;
-	private int mousestartlocationx = -1, mousestartlocationy = -1;  
-	private int mouselastlocationx = -1, mouselastlocationy = -1;  
-	private int mouselocationx = -1, mouselocationy = -1;
-	private boolean drawlinemode = false;
-	private JFileChooser filechooser = new JFileChooser();
-	private ImageFileFilters.PNGFileFilter pngfilefilter = new ImageFileFilters.PNGFileFilter();
-	private ImageFileFilters.JPGFileFilter jpgfilefilter = new ImageFileFilters.JPGFileFilter();
-	private ImageFileFilters.GIFFileFilter giffilefilter = new ImageFileFilters.GIFFileFilter();
-	private ImageFileFilters.BMPFileFilter bmpfilefilter = new ImageFileFilters.BMPFileFilter();
-	private ImageFileFilters.WBMPFileFilter wbmpfilefilter = new ImageFileFilters.WBMPFileFilter();
-	private DragAndDropClipBoardHandler dndcbhandler = new DragAndDropClipBoardHandler();
 	private DropTargetHandler droptargethandler = new DropTargetHandler();
 	
 	public JavaRenderEngine() {
@@ -96,23 +57,17 @@ public class JavaRenderEngine extends JFrame implements KeyListener,MouseListene
 		}else {
 			this.setLocationByPlatform(true);
 		}
-		this.filechooser.addChoosableFileFilter(this.pngfilefilter);
-		this.filechooser.addChoosableFileFilter(this.jpgfilefilter);
-		this.filechooser.addChoosableFileFilter(this.giffilefilter);
-		this.filechooser.addChoosableFileFilter(this.bmpfilefilter);
-		this.filechooser.addChoosableFileFilter(this.wbmpfilefilter);
-		this.filechooser.setFileFilter(pngfilefilter);
 		this.addKeyListener(this);
 		this.renderpanel.addMouseListener(this);
 		this.renderpanel.addMouseMotionListener(this);
 		this.renderpanel.addMouseWheelListener(this);
-		this.renderpanel.setTransferHandler(dndcbhandler);
 		this.renderpanel.setDropTarget(droptargethandler);
 		this.renderpanel.setPreferredSize(new Dimension(this.imagecanvaswidth,this.imagecanvasheight));
 		this.renderpanel.setSize(this.imagecanvaswidth,this.imagecanvasheight);
 		this.setContentPane(renderpanel);
 		this.pack();
 		this.setVisible(true);
+		this.setActiveApp(drawapp);
 	}
 
 	public static void main(String[] args) {
@@ -198,24 +153,8 @@ public class JavaRenderEngine extends JFrame implements KeyListener,MouseListene
 		private final int fpstargetdelay = (int)Math.floor(1000.0f/(2.0f*(double)fpstarget));
 		private final Timer timer = new Timer(fpstargetdelay,this);
 		private long lastupdate = System.currentTimeMillis();
-		private VolatileImage renderbuffer = null;
-		private VolatileImage dragbuffer = null;
-		private TexturePaint bgpattern = null;
-		public RenderPanel(int imagewidth, int imageheight) {
-			this.renderbuffer = gc.createCompatibleVolatileImage(imagewidth, imageheight, Transparency.TRANSLUCENT);
-			this.dragbuffer = gc.createCompatibleVolatileImage(imagewidth,imageheight,Transparency.TRANSLUCENT);
-			Graphics2D gfx = this.renderbuffer.createGraphics();
-			gfx.setComposite(AlphaComposite.Clear);
-			gfx.fillRect(0, 0, imagewidth, imageheight);
-			BufferedImage bgpatternimage = gc.createCompatibleImage(64, 64, Transparency.OPAQUE);
-			Graphics2D pgfx = bgpatternimage.createGraphics();
-			pgfx.setColor(Color.WHITE);
-			pgfx.fillRect(0, 0, bgpatternimage.getWidth(), bgpatternimage.getHeight());
-			pgfx.setColor(Color.BLACK);
-			pgfx.drawLine(31, 0, 31, 63);
-			pgfx.drawLine(0, 31, 63, 31);
-			pgfx.dispose();
-			this.bgpattern = new TexturePaint(bgpatternimage,new Rectangle(0, 0, 64, 64));
+		private VolatileImage doublebuffer = null;
+		public RenderPanel() {
 			this.addComponentListener(this);
 			timer.start();
 		}
@@ -224,14 +163,15 @@ public class JavaRenderEngine extends JFrame implements KeyListener,MouseListene
 			super.paintComponent(g);
 			long newupdate = System.currentTimeMillis();
 			long ticktime = newupdate-lastupdate;
+			double ticktimesec = (double)ticktime / 1000.0f;
 			double ticktimefps = 1000.0f/(double)ticktime;
 			lastupdate = newupdate;
-			if (JavaRenderEngine.this.resizing) {
+			if ((doublebuffer==null)||((JavaRenderEngine.this.resizing)&&(doublebuffer.getWidth()!=this.getWidth())&&(doublebuffer.getHeight()!=this.getHeight()))) {
 				JavaRenderEngine.this.resizing = false;
-				VolatileImage oldimage = this.renderbuffer;
-				this.renderbuffer = gc.createCompatibleVolatileImage(this.getWidth(),this.getHeight(), Transparency.TRANSLUCENT);
-				this.dragbuffer = gc.createCompatibleVolatileImage(this.getWidth(),this.getHeight(), Transparency.TRANSLUCENT);
-				Graphics2D gfx = this.renderbuffer.createGraphics();
+				System.out.println("Window: Resolution "+this.getWidth()+"x"+this.getHeight());
+				VolatileImage oldimage = this.doublebuffer;
+				this.doublebuffer = gc.createCompatibleVolatileImage(this.getWidth(),this.getHeight(), Transparency.OPAQUE);
+				Graphics2D gfx = this.doublebuffer.createGraphics();
 				gfx.setComposite(AlphaComposite.Clear);
 				gfx.fillRect(0, 0, this.getWidth(),this.getHeight());
 				if (oldimage!=null) {
@@ -240,562 +180,82 @@ public class JavaRenderEngine extends JFrame implements KeyListener,MouseListene
 				}
 				gfx.dispose();
 			}
-			Graphics2D g2 = (Graphics2D)g;
-			if (renderbuffer!=null) {
-				g2.setPaint(this.bgpattern);
-				g2.fillRect(0, 0, renderbuffer.getWidth(), renderbuffer.getHeight());
-				g2.setPaint(null);
-				g2.drawImage(renderbuffer, 0, 0, null);
-	    		if ((JavaRenderEngine.this.penciloverridemode)&&(JavaRenderEngine.this.pencilbuffer!=null)) {
-	        		double pencilsizescalefactor = ((double)JavaRenderEngine.this.pencilsize)/((double)JavaRenderEngine.this.pencilbuffer.getWidth());
-	    			g2.setComposite(AlphaComposite.Src);
-					g2.setPaint(this.bgpattern);
-					g2.fillRect(JavaRenderEngine.this.mouselocationx-(int)Math.round((double)JavaRenderEngine.this.pencilbuffer.getWidth()*pencilsizescalefactor/2.0f), JavaRenderEngine.this.mouselocationy-(int)Math.round((double)JavaRenderEngine.this.pencilbuffer.getHeight()*pencilsizescalefactor/2.0f),(int)Math.round(JavaRenderEngine.this.pencilbuffer.getWidth()*pencilsizescalefactor),(int)Math.round(JavaRenderEngine.this.pencilbuffer.getHeight()*pencilsizescalefactor));
-	    		}
-    			if (JavaRenderEngine.this.drawlinemode) {
-    				JavaRenderEngine.this.drawPencilLine(g2, JavaRenderEngine.this.mousestartlocationx, JavaRenderEngine.this.mousestartlocationy, JavaRenderEngine.this.mouselocationx, JavaRenderEngine.this.mouselocationy, false, false);
-    			} else {
-    				JavaRenderEngine.this.drawPencil(g2, JavaRenderEngine.this.mouselocationx, JavaRenderEngine.this.mouselocationy, false, false);
+			if (this.doublebuffer!=null) {
+				Graphics2D doublebuffergfx = this.doublebuffer.createGraphics();
+				if (JavaRenderEngine.this.activeapp!=null) {
+					JavaRenderEngine.this.activeapp.renderWindow(doublebuffergfx, doublebuffer.getWidth(), doublebuffer.getHeight(), JavaRenderEngine.this.resizing, ticktimesec, ticktimefps);
 				}
+				Graphics2D g2 = (Graphics2D)g;
+				g2.drawImage(this.doublebuffer,0,0,null);
 			}
 		}
 		@Override
 		public void actionPerformed(ActionEvent e) {this.repaint();}
-		public VolatileImage getRenderBuffer() {return renderbuffer;}
-		public VolatileImage getDragBuffer() {return dragbuffer;}
-		public void setRenderBuffer(VolatileImage renderbufferi) {this.renderbuffer=renderbufferi;}
 		
 		@Override public void componentMoved(ComponentEvent e) {}
 		@Override public void componentShown(ComponentEvent e) {}
 		@Override public void componentHidden(ComponentEvent e) {}
 		@Override public void componentResized(ComponentEvent e) {
-			JavaRenderEngine.this.resizewidth = this.getWidth();
-			JavaRenderEngine.this.resizeheight = this.getHeight();
 			JavaRenderEngine.this.resizing = true;
+			System.out.println("Window: Resizing");
 		}
 	}
 	
-	private void setActiveApp(AppHandler activeappi) {this.activeapp = activeappi;}
-	
-	private interface AppHandler extends ActionListener,ComponentListener,KeyListener,MouseListener,MouseMotionListener,MouseWheelListener {}
-	private class DrawApp implements AppHandler {
-
-		@Override
-		public void actionPerformed(ActionEvent e) {
-		}
-
-		@Override
-		public void componentResized(ComponentEvent e) {
-		}
-
-		@Override
-		public void componentMoved(ComponentEvent e) {
-		}
-
-		@Override
-		public void componentShown(ComponentEvent e) {
-		}
-
-		@Override
-		public void componentHidden(ComponentEvent e) {
-		}
-
-		@Override
-		public void keyTyped(KeyEvent e) {
-		}
-
-		@Override
-		public void keyPressed(KeyEvent e) {
-		}
-
-		@Override
-		public void keyReleased(KeyEvent e) {
-		}
-
-		@Override
-		public void mouseClicked(MouseEvent e) {
-		}
-
-		@Override
-		public void mousePressed(MouseEvent e) {
-		}
-
-		@Override
-		public void mouseReleased(MouseEvent e) {
-		}
-
-		@Override
-		public void mouseEntered(MouseEvent e) {
-		}
-
-		@Override
-		public void mouseExited(MouseEvent e) {
-		}
-
-		@Override
-		public void mouseDragged(MouseEvent e) {
-		}
-
-		@Override
-		public void mouseMoved(MouseEvent e) {
-		}
-
-		@Override
-		public void mouseWheelMoved(MouseWheelEvent e) {
-		}
-	}
-	
-	private class ImageFileFilters  {
-		public static class PNGFileFilter extends FileFilter {
-			@Override public boolean accept(File f) {return (f.isDirectory())||(f.getName().endsWith(".png"));}
-			@Override public String getDescription() {return "PNG Image file";}
-		}
-		public static class JPGFileFilter extends FileFilter {
-			@Override public boolean accept(File f) {return (f.isDirectory())||(f.getName().endsWith(".jpg"))||(f.getName().endsWith(".jpeg"));}
-			@Override public String getDescription() {return "JPG Image file";}
-		}
-		public static class GIFFileFilter extends FileFilter {
-			@Override public boolean accept(File f) {return (f.isDirectory())||(f.getName().endsWith(".gif"));}
-			@Override public String getDescription() {return "GIF Image file";}
-		}
-		public static class BMPFileFilter extends FileFilter {
-			@Override public boolean accept(File f) {return (f.isDirectory())||(f.getName().endsWith(".bmp"));}
-			@Override public String getDescription() {return "BMP Image file";}
-		}
-		public static class WBMPFileFilter extends FileFilter {
-			@Override public boolean accept(File f) {return (f.isDirectory())||(f.getName().endsWith(".wbmp"));}
-			@Override public String getDescription() {return "WBMP Image file";}
-		}
-	}
-	private class DragAndDropClipBoardHandler extends TransferHandler {
-		private static final long serialVersionUID = 1L;
-        public boolean canImport(TransferHandler.TransferSupport info) {System.out.println("canImport");return false;}
-        public boolean importData(TransferHandler.TransferSupport info) {System.out.println("importData");return false;}
-        public int getSourceActions(JComponent c) {System.out.println("getSourceActions");return COPY;}
-        protected Transferable createTransferable(JComponent c) {System.out.println("createTransferable");return null;}
-	}
 	private class DropTargetHandler extends DropTarget {
 		private static final long serialVersionUID = 1L;
 		@Override public synchronized void drop(DropTargetDropEvent dtde) {System.out.println("drop");dtde.rejectDrop();}
 	}
 
-	private void drawPencil(Graphics2D g, int mousex, int mousey, boolean erasemode, boolean overridemode) {
-		g.setComposite(AlphaComposite.SrcOver);
-		g.setPaint(null);
-		g.setColor(null);
-		int pencilwidth = (int)Math.ceil((double)(this.pencilsize-1)/2.0f);
-    	if (this.pencilbuffer!=null) {
-	    	if (erasemode) {
-	    		if (overridemode) {
-	    			g.setComposite(AlphaComposite.Clear);
-	    		} else {
-	    			g.setComposite(AlphaComposite.DstOut);
-	    		}
-	    	} else {
-	    		if (overridemode) {
-	    			g.setComposite(AlphaComposite.Src);
-	    		}
-    		}
-    		double pencilsizescalefactor = ((double)this.pencilsize)/((double)this.pencilbuffer.getWidth());
-    		g.drawImage(JavaRenderEngine.this.pencilbuffer, mousex-(int)Math.round((double)JavaRenderEngine.this.pencilbuffer.getWidth()*pencilsizescalefactor/2.0f), mousey-(int)Math.round((double)JavaRenderEngine.this.pencilbuffer.getHeight()*pencilsizescalefactor/2.0f),(int)Math.round(JavaRenderEngine.this.pencilbuffer.getWidth()*pencilsizescalefactor),(int)Math.round(JavaRenderEngine.this.pencilbuffer.getHeight()*pencilsizescalefactor),null);
-    	} else {
-	    	if (erasemode) {
-	    		g.setComposite(AlphaComposite.Src);
-	    		g.setColor(this.erasecolor);
-	    	} else {
-	    		if (overridemode) {
-		    		g.setComposite(AlphaComposite.Src);
-	    		}
-    			g.setColor(this.drawcolor);
-	    	}
-			if (this.pencilshape==2) {
-				g.fillRoundRect(mousex-pencilwidth, mousey-pencilwidth, this.pencilsize, this.pencilsize, 5, 5);
-			} else if (this.pencilshape==3) {
-				g.fillOval(mousex-pencilwidth, mousey-pencilwidth, this.pencilsize, this.pencilsize);
-			} else if (this.pencilshape==4) {
-				g.drawRect(mousex-pencilwidth, mousey-pencilwidth, this.pencilsize, this.pencilsize);
-			} else if (this.pencilshape==5) {
-				g.drawRoundRect(mousex-pencilwidth, mousey-pencilwidth, this.pencilsize, this.pencilsize, 5, 5);
-			} else if (this.pencilshape==6) {
-				g.drawOval(mousex-pencilwidth, mousey-pencilwidth, this.pencilsize, this.pencilsize);
-			}else {
-				g.fillRect(mousex-pencilwidth, mousey-pencilwidth, this.pencilsize, this.pencilsize);
-			}
-    	}
+	private void setActiveApp(AppHandler activeappi) {
+		//this.activeapp = activeappi;
 	}
-	private void drawPencilLine(Graphics2D g, int mousestartx, int mousestarty, int mousex, int mousey, boolean erasemode, boolean overridemode) {
-		double linedistx = mousex-mousestartx;
-		double linedisty = mousey-mousestarty;
-		int linestepnum = (int)Math.ceil(Math.sqrt(linedistx*linedistx+linedisty*linedisty))+1;
-		double linestepx = linedistx/linestepnum;
-		double linestepy = linedisty/linestepnum;
-		for (int i=0;i<linestepnum;i++) {
-			int drawposx = (int)Math.round(this.mousestartlocationx + i*linestepx);
-			int drawposy = (int)Math.round(this.mousestartlocationy + i*linestepy);
-	    	this.drawPencil(g, drawposx, drawposy, erasemode, overridemode);
-		}
+	public interface AppHandler extends ActionListener,ComponentListener,KeyListener,MouseListener,MouseMotionListener,MouseWheelListener {
+		public void renderWindow(Graphics2D g, int renderwidth, int renderheight, boolean resizing, double deltatimesec, double deltatimefps);
 	}
-	
-	@Override public void keyTyped(KeyEvent e) {}
-	@Override public void keyReleased(KeyEvent e) {}
 
-	@Override
-	public void keyPressed(KeyEvent e) {
+	@Override public void mouseWheelMoved(MouseWheelEvent e) {if (this.activeapp!=null) {this.activeapp.mouseWheelMoved(e);}}
+	@Override public void mouseDragged(MouseEvent e) {if (this.activeapp!=null) {this.activeapp.mouseDragged(e);}}
+	@Override public void mouseMoved(MouseEvent e) {if (this.activeapp!=null) {this.activeapp.mouseMoved(e);}}
+	@Override public void mouseClicked(MouseEvent e) {if (this.activeapp!=null) {this.activeapp.mouseClicked(e);}}
+	@Override public void mousePressed(MouseEvent e) {if (this.activeapp!=null) {this.activeapp.mousePressed(e);}}
+	@Override public void mouseReleased(MouseEvent e) {if (this.activeapp!=null) {this.activeapp.mouseReleased(e);}}
+	@Override public void mouseEntered(MouseEvent e) {if (this.activeapp!=null) {this.activeapp.mouseEntered(e);}}
+	@Override public void mouseExited(MouseEvent e) {if (this.activeapp!=null) {this.activeapp.mouseExited(e);}}
+	@Override public void keyTyped(KeyEvent e) {if (this.activeapp!=null) {this.activeapp.keyTyped(e);}}
+	@Override public void keyReleased(KeyEvent e) {if (this.activeapp!=null) {this.activeapp.keyReleased(e);}}
+
+	@Override public void keyPressed(KeyEvent e) {
+	    int altonmask = KeyEvent.ALT_DOWN_MASK;
+	    int altoffmask = KeyEvent.SHIFT_DOWN_MASK|KeyEvent.CTRL_DOWN_MASK;
+	    boolean altdownmask = (e.getModifiersEx() & (altonmask | altoffmask)) == altonmask;
+	    
 		if (e.getKeyCode()==KeyEvent.VK_ESCAPE) {
+			System.out.println("keyPressed: VK_ESCAPE");
 			System.exit(0);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_ENTER) {
-		    int onmask = KeyEvent.ALT_DOWN_MASK;
-		    int offmask = KeyEvent.SHIFT_DOWN_MASK|KeyEvent.CTRL_DOWN_MASK;
-		    boolean enteraltdown = (e.getModifiersEx() & (onmask | offmask)) == onmask;
-		    if (enteraltdown) {
-		    	JavaRenderEngine.this.dispose();
-		    	if (!windowedmode) {
-		    		windowedmode = true;
-		    		JavaRenderEngine.this.setExtendedState(JavaRenderEngine.this.getExtendedState()&~JFrame.MAXIMIZED_BOTH);
-		    		JavaRenderEngine.this.setUndecorated(false);
-		    	}else {
-		    		windowedmode = false;
-		    		JavaRenderEngine.this.setExtendedState(JavaRenderEngine.this.getExtendedState()|JFrame.MAXIMIZED_BOTH);
-		    		JavaRenderEngine.this.setUndecorated(true);
-		    	}
-		    	JavaRenderEngine.this.setVisible(true);
-		    }
-		    int onmaska = 0;
-		    int offmaska = KeyEvent.ALT_DOWN_MASK|KeyEvent.SHIFT_DOWN_MASK|KeyEvent.CTRL_DOWN_MASK;
-		    boolean enterdown = (e.getModifiersEx() & (onmaska | offmaska)) == onmaska;
-		    if(enterdown) {
-		    	this.penciloverridemode = !this.penciloverridemode;
-		    }
-		}
-		if (e.getKeyCode()==KeyEvent.VK_BACK_SPACE) {
-			VolatileImage renderbufferhandle = renderpanel.getRenderBuffer();
-			if (renderbufferhandle!=null) {
-				Graphics2D gfx = renderbufferhandle.createGraphics();
-				gfx.setComposite(AlphaComposite.Src);
-				gfx.setColor(this.erasecolor);
-				gfx.fillRect(0, 0, renderbufferhandle.getWidth(), renderbufferhandle.getHeight());
-				gfx.dispose();
-			}
-		}
-		if (e.getKeyCode()==KeyEvent.VK_INSERT) {
-			this.drawcolorhsb[0] += 0.01f;
-			if (this.drawcolorhsb[0]>1.0f) {this.drawcolorhsb[0] = 0.0f;}
-			Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-			float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-			this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_DELETE) {
-			this.drawcolorhsb[0] -= 0.01f;
-			if (this.drawcolorhsb[0]<0.0f) {this.drawcolorhsb[0] = 1.0f;}
-			Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-			float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-			this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_HOME) {
-			this.drawcolorhsb[1] += 0.01f;
-			if (this.drawcolorhsb[1]>1.0f) {this.drawcolorhsb[1] = 1.0f;}
-			Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-			float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-			this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_END) {
-			this.drawcolorhsb[1] -= 0.01f;
-			if (this.drawcolorhsb[1]<0.0f) {this.drawcolorhsb[1] = 0.0f;}
-			Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-			float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-			this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_PAGE_UP) {
-			this.drawcolorhsb[2] += 0.01f;
-			if (this.drawcolorhsb[2]>1.0f) {this.drawcolorhsb[2] = 1.0f;}
-			Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-			float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-			this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_PAGE_DOWN) {
-			this.drawcolorhsb[2] -= 0.01f;
-			if (this.drawcolorhsb[2]<0.0f) {this.drawcolorhsb[2] = 0.0f;}
-			Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-			float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-			this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_ADD) {
-			this.pencilsize += 1;
-		}
-		if (e.getKeyCode()==KeyEvent.VK_SUBTRACT) {
-			this.pencilsize -= 1;
-			if (this.pencilsize<1) {this.pencilsize = 1;}
-		}
-		if (e.getKeyCode()==KeyEvent.VK_DIVIDE) {
-	    	this.pencilshape -= 1;
-	    	if (this.pencilshape<1) {this.pencilshape = 6;}
-	    	if (this.pencilbuffer!=null) {
-	    		this.pencilbuffer = null;
-	    		this.pencilsize = this.oldpencilsize;
+		}else if ((e.getKeyCode()==KeyEvent.VK_ENTER)&&(altdownmask)) {
+			System.out.println("keyPressed: ALT+VK_ENTER");
+	    	JavaRenderEngine.this.dispose();
+	    	if (!windowedmode) {
+	    		windowedmode = true;
+	    		JavaRenderEngine.this.setExtendedState(JavaRenderEngine.this.getExtendedState()&~JFrame.MAXIMIZED_BOTH);
+	    		JavaRenderEngine.this.setUndecorated(false);
+	    	}else {
+	    		windowedmode = false;
+	    		JavaRenderEngine.this.setExtendedState(JavaRenderEngine.this.getExtendedState()|JFrame.MAXIMIZED_BOTH);
+	    		JavaRenderEngine.this.setUndecorated(true);
 	    	}
+	    	JavaRenderEngine.this.setVisible(true);
+		}else if (e.getKeyCode()==KeyEvent.VK_F5) {
+			System.out.println("keyPressed: VK_F5");
+			this.setActiveApp(drawapp);
+		}else if (e.getKeyCode()==KeyEvent.VK_F6) {
+			System.out.println("keyPressed: VK_F6");
+		}else if (e.getKeyCode()==KeyEvent.VK_F7) {
+			System.out.println("keyPressed: VK_F7");
+		}else if (e.getKeyCode()==KeyEvent.VK_F8) {
+			System.out.println("keyPressed: VK_F8");
+		}else {
+			if (this.activeapp!=null) {this.activeapp.keyPressed(e);}
 		}
-		if (e.getKeyCode()==KeyEvent.VK_MULTIPLY) {
-	    	this.pencilshape += 1;
-	    	if (this.pencilshape>6) {this.pencilshape = 1;}
-	    	if (this.pencilbuffer!=null) {
-	    		this.pencilbuffer = null;
-	    		this.pencilsize = this.oldpencilsize;
-	    	}
-		}
-		if (e.getKeyCode()==KeyEvent.VK_NUMPAD9) {
-			this.penciltransparency += 0.01f;
-			if (this.penciltransparency>1.0f) {this.penciltransparency = 1.0f;}
-			float[] colorvalues = this.drawcolor.getRGBColorComponents(new float[3]);
-			this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_NUMPAD8) {
-			this.penciltransparency -= 0.01f;
-			if (this.penciltransparency<0.0f) {this.penciltransparency = 0.0f;}
-			float[] colorvalues = this.drawcolor.getRGBColorComponents(new float[3]);
-			this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F1) {
-			//TODO help pop-up window
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F2) {
-			this.filechooser.setDialogTitle("Save File");
-			this.filechooser.setApproveButtonText("Save");
-			if (this.filechooser.showOpenDialog(this)==JFileChooser.APPROVE_OPTION) {
-				File savefile = this.filechooser.getSelectedFile();
-				FileFilter savefileformat = this.filechooser.getFileFilter();
-				if (savefileformat.equals(this.jpgfilefilter)) {
-					try {ImageIO.write(renderpanel.getRenderBuffer().getSnapshot(), "JPG", savefile);} catch (Exception ex) {ex.printStackTrace();}
-				} else if (savefileformat.equals(this.giffilefilter)) {
-					try {ImageIO.write(renderpanel.getRenderBuffer().getSnapshot(), "GIF", savefile);} catch (Exception ex) {ex.printStackTrace();}
-				} else if (savefileformat.equals(this.bmpfilefilter)) {
-					try {ImageIO.write(renderpanel.getRenderBuffer().getSnapshot(), "BMP", savefile);} catch (Exception ex) {ex.printStackTrace();}
-				} else if (savefileformat.equals(this.wbmpfilefilter)) {
-					try {ImageIO.write(renderpanel.getRenderBuffer().getSnapshot(), "WBMP", savefile);} catch (Exception ex) {ex.printStackTrace();}
-				} else {
-					try {ImageIO.write(renderpanel.getRenderBuffer().getSnapshot(), "PNG", savefile);} catch (Exception ex) {ex.printStackTrace();}
-				}
-			}
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F3) {
-			this.filechooser.setDialogTitle("Load File");
-			this.filechooser.setApproveButtonText("Load");
-			if (this.filechooser.showOpenDialog(this)==JFileChooser.APPROVE_OPTION) {
-				File loadfile = this.filechooser.getSelectedFile();
-				BufferedImage loadimage = null;
-				try {loadimage=ImageIO.read(loadfile);} catch (Exception ex) {ex.printStackTrace();}
-				if (loadimage!=null) {
-				    int onmask = KeyEvent.SHIFT_DOWN_MASK;
-				    int offmask = KeyEvent.ALT_DOWN_MASK|KeyEvent.CTRL_DOWN_MASK;
-				    boolean f3shiftdown = (e.getModifiersEx() & (onmask | offmask)) == onmask;
-				    if (f3shiftdown) {
-				    	this.oldpencilsize = this.pencilsize;
-						this.pencilsize = loadimage.getWidth();
-						VolatileImage loadimagevolatile = gc.createCompatibleVolatileImage(loadimage.getWidth(), loadimage.getHeight(), Transparency.TRANSLUCENT);
-						Graphics2D loadimagevolatilegfx = loadimagevolatile.createGraphics();
-						loadimagevolatilegfx.setComposite(AlphaComposite.Src);
-						loadimagevolatilegfx.drawImage(loadimage, 0, 0, null);
-						loadimagevolatilegfx.dispose();
-				    	this.pencilbuffer = loadimagevolatile;
-				    	this.pencilbuffer.setAccelerationPriority(1.0f);
-				    }else{
-						if (this.windowedmode) {
-							VolatileImage loadimagevolatile = gc.createCompatibleVolatileImage(loadimage.getWidth(), loadimage.getHeight(), Transparency.TRANSLUCENT);
-							Graphics2D loadimagevolatilegfx = loadimagevolatile.createGraphics();
-							loadimagevolatilegfx.setComposite(AlphaComposite.Src);
-							loadimagevolatilegfx.drawImage(loadimage, 0, 0, null);
-							loadimagevolatilegfx.dispose();
-					    	this.renderpanel.setRenderBuffer(loadimagevolatile);
-							this.renderpanel.setPreferredSize(new Dimension(loadimage.getWidth(),loadimage.getHeight()));
-							this.renderpanel.setSize(loadimage.getWidth(),loadimage.getHeight());
-							this.pack();
-						}else {
-							VolatileImage loadimagevolatile = gc.createCompatibleVolatileImage(this.renderpanel.getWidth(), this.renderpanel.getHeight(), Transparency.TRANSLUCENT);
-							Graphics2D loadimagevolatilegfx = loadimagevolatile.createGraphics();
-							loadimagevolatilegfx.setComposite(AlphaComposite.Clear);
-							loadimagevolatilegfx.fillRect(0, 0, this.renderpanel.getWidth(), this.renderpanel.getHeight());
-							loadimagevolatilegfx.setComposite(AlphaComposite.Src);
-							loadimagevolatilegfx.drawImage(loadimage, 0, 0, null);
-							loadimagevolatilegfx.dispose();
-					    	this.renderpanel.setRenderBuffer(loadimagevolatile);
-						}
-				    }
-				}
-			}
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F4) {
-			//TODO tools/color pop-up window
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F5) {
-			//TODO 2D image edit mode
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F6) {
-			//TODO CAD 2D edit mode
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F7) {
-			//TODO CAD 3D edit mode
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F8) {
-			//TODO Java code edit mode
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F9) {
-			//TODO Game run mode
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F10) {
-			//TODO <tbd>
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F11) {
-			//TODO <tbd>
-		}
-		if (e.getKeyCode()==KeyEvent.VK_F12) {
-			//TODO Save screen shot
-		}
-	}
-
-	@Override public void mouseClicked(MouseEvent e) {}
-	@Override public void mouseEntered(MouseEvent e) {}
-	@Override public void mouseExited(MouseEvent e) {}
-	
-	@Override public void mouseMoved(MouseEvent e) {this.mouselocationx=e.getX();this.mouselocationy=e.getY();}
-	@Override public void mousePressed(MouseEvent e) {this.mouselocationx=e.getX();this.mouselocationy=e.getY();this.mousestartlocationx=this.mouselocationx;this.mousestartlocationy=this.mouselocationy;mouseDragged(e);}
-	@Override public void mouseReleased(MouseEvent e) {
-		VolatileImage renderbufferhandle = this.renderpanel.getRenderBuffer();
-		if (renderbufferhandle!=null) {
-			Graphics2D renderbuffergfx = renderbufferhandle.createGraphics();
-			renderbuffergfx.setColor(JavaRenderEngine.this.drawcolor);
-		    boolean mouse1up = e.getButton()==MouseEvent.BUTTON1;
-		    boolean mouse3up = e.getButton()==MouseEvent.BUTTON3;
-			if (mouse1up||mouse3up) {
-				if (this.drawlinemode) {
-					this.drawlinemode=false;
-					drawPencilLine(renderbuffergfx, this.mousestartlocationx, this.mousestartlocationy, this.mouselocationx, this.mouselocationy, mouse3up, this.penciloverridemode);
-			    	renderbufferhandle.contentsLost();
-				}
-			}
-			renderbuffergfx.dispose();
-		}
-	}
-	
-	@Override
-	public void mouseDragged(MouseEvent e) {
-		this.mouselastlocationx=this.mouselocationx;this.mouselastlocationy=this.mouselocationy;
-		this.mouselocationx=e.getX();this.mouselocationy=e.getY();
-    	int mousedeltax = this.mouselocationx - this.mouselastlocationx; 
-    	int mousedeltay = this.mouselocationy - this.mouselastlocationy;
-		VolatileImage renderbufferhandle = renderpanel.getRenderBuffer();
-		if (renderbufferhandle!=null) {
-			Graphics2D renderbuffergfx = renderbufferhandle.createGraphics();
-		    int onmask1 = MouseEvent.BUTTON1_DOWN_MASK;
-		    int offmask1 = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-		    int onmask3 = MouseEvent.BUTTON3_DOWN_MASK;
-		    int offmask3 = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-		    boolean mouse1down = ((e.getModifiersEx() & (onmask1 | offmask1)) == onmask1);
-		    boolean mouse3down = ((e.getModifiersEx() & (onmask3 | offmask3)) == onmask3);
-		    if (mouse1down||mouse3down) {
-		    	this.drawPencil(renderbuffergfx, e.getX(), e.getY(), mouse3down, this.penciloverridemode);
-		    	renderbufferhandle.contentsLost();
-			}			
-		    int onmask1c = MouseEvent.BUTTON1_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-		    int offmask1c = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK;
-		    int onmask3c = MouseEvent.BUTTON3_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-		    int offmask3c = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK;
-		    boolean mouse1altdown = ((e.getModifiersEx() & (onmask1c | offmask1c)) == onmask1c);
-		    boolean mouse3altdown = ((e.getModifiersEx() & (onmask3c | offmask3c)) == onmask3c);
-		    if (mouse1altdown||mouse3altdown) {
-		    	this.drawlinemode = true;
-		    }
-		    int onmask1a = MouseEvent.BUTTON1_DOWN_MASK|MouseEvent.SHIFT_DOWN_MASK;
-		    int offmask1a = MouseEvent.CTRL_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-		    boolean mouse1shiftdown = ((e.getModifiersEx() & (onmask1a | offmask1a)) == onmask1a);
-		    if (mouse1shiftdown) {
-				int colorvalue = renderbufferhandle.getSnapshot().getRGB(e.getX(), e.getY());
-				Color pickeddrawcolor = new Color(colorvalue);
-				this.drawcolorhsb = Color.RGBtoHSB(pickeddrawcolor.getRed(), pickeddrawcolor.getGreen(), pickeddrawcolor.getBlue(), new float[3]);
-				float[] colorvalues = pickeddrawcolor.getRGBColorComponents(new float[3]);
-				this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-		    }
-		    int onmask1b = MouseEvent.BUTTON1_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK;
-		    int offmask1b = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-		    boolean mouse1controldown = ((e.getModifiersEx() & (onmask1b | offmask1b)) == onmask1b);
-		    if (mouse1controldown) {
-		    	//TODO select canvas region
-		    }
-		    int onmask2 = MouseEvent.BUTTON2_DOWN_MASK;
-		    int offmask2 = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-		    boolean mouse2down = ((e.getModifiersEx() & (onmask2 | offmask2)) == onmask2);
-		    if (mouse2down) {
-		    	//TODO <tbd>
-		    }
-		    int onmask2a = MouseEvent.BUTTON2_DOWN_MASK|MouseEvent.SHIFT_DOWN_MASK;
-		    int offmask2a = MouseEvent.CTRL_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-		    boolean mouse2shiftdown = ((e.getModifiersEx() & (onmask2a | offmask2a)) == onmask2a);
-		    if (mouse2shiftdown) {
-		    	VolatileImage dragimage = JavaRenderEngine.this.renderpanel.getDragBuffer();
-		    	Graphics2D dragimagegfx = dragimage.createGraphics();
-		    	dragimagegfx.setComposite(AlphaComposite.Clear);
-		    	dragimagegfx.fillRect(0, 0, renderbufferhandle.getWidth(), renderbufferhandle.getHeight());
-		    	dragimagegfx.setComposite(AlphaComposite.Src);
-		    	dragimagegfx.drawImage(renderbufferhandle, mousedeltax, mousedeltay, null);
-		    	dragimagegfx.dispose();
-		    	renderbuffergfx.setComposite(AlphaComposite.Clear);
-		    	renderbuffergfx.fillRect(0, 0, renderbufferhandle.getWidth(), renderbufferhandle.getHeight());
-		    	renderbuffergfx.setComposite(AlphaComposite.Src);
-		    	renderbuffergfx.drawImage(dragimage, 0, 0, null);
-		    	renderbufferhandle.contentsLost();
-		    }
-	    	renderbuffergfx.dispose();
-		}
-	}
-	
-	@Override
-	public void mouseWheelMoved(MouseWheelEvent e) {
-	    int onmask4 = 0;
-	    int offmask4 = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-	    boolean mousewheeldown = ((e.getModifiersEx() & (onmask4 | offmask4)) == onmask4);
-	    if (mousewheeldown) {
-			this.pencilsize += e.getWheelRotation();
-			if (this.pencilsize<1) {
-				this.pencilsize = 1;
-			}
-	    }
-	    int onmask4a = MouseEvent.CTRL_DOWN_MASK;
-	    int offmask4a = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-	    boolean mousewheelctrldown = ((e.getModifiersEx() & (onmask4a | offmask4a)) == onmask4a);
-	    if (mousewheelctrldown) {
-	    	this.drawcolorhsb[0] += 0.01f*e.getWheelRotation();
-	    	if (this.drawcolorhsb[0]>1.0f) {this.drawcolorhsb[0] = 0.0f;}
-	    	else if (this.drawcolorhsb[0]<0.0f) {this.drawcolorhsb[0] = 1.0f;}
-	    	Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-	    	float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-	    	this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-	    }
-	    int onmask4b = MouseEvent.ALT_DOWN_MASK;
-	    int offmask4b = MouseEvent.SHIFT_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK;
-	    boolean mousewheelaltdown = ((e.getModifiersEx() & (onmask4b | offmask4b)) == onmask4b);
-	    if (mousewheelaltdown) {
-	    	this.drawcolorhsb[2] += 0.01f*e.getWheelRotation();
-	    	if (this.drawcolorhsb[2]>1.0f) {this.drawcolorhsb[2] = 1.0f;}
-	    	else if (this.drawcolorhsb[2]<0.0f) {this.drawcolorhsb[2] = 0.0f;}
-	    	Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-	    	float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-	    	this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-	    }
-	    int onmask4c = MouseEvent.ALT_DOWN_MASK|MouseEvent.CTRL_DOWN_MASK;
-	    int offmask4c = MouseEvent.SHIFT_DOWN_MASK;
-	    boolean mousewheelctrlaltdown = ((e.getModifiersEx() & (onmask4c | offmask4c)) == onmask4c);
-	    if (mousewheelctrlaltdown) {
-	    	this.drawcolorhsb[1] += 0.01f*e.getWheelRotation();
-	    	if (this.drawcolorhsb[1]>1.0f) {this.drawcolorhsb[1] = 1.0f;}
-	    	else if (this.drawcolorhsb[1]<0.0f) {this.drawcolorhsb[1] = 0.0f;}
-	    	Color hsbcolor = Color.getHSBColor(this.drawcolorhsb[0], this.drawcolorhsb[1], this.drawcolorhsb[2]);
-	    	float[] colorvalues = hsbcolor.getRGBColorComponents(new float[3]);
-	    	this.drawcolor = new Color(colorvalues[0],colorvalues[1],colorvalues[2],this.penciltransparency);
-	    }
-	    int onmask4d = MouseEvent.SHIFT_DOWN_MASK;
-	    int offmask4d = MouseEvent.CTRL_DOWN_MASK|MouseEvent.ALT_DOWN_MASK;
-	    boolean mousewheelshiftdown = ((e.getModifiersEx() & (onmask4d | offmask4d)) == onmask4d);
-	    if (mousewheelshiftdown) {
-	    	//TODO <tbd>
-	    }
 	}
 }
